Consolidate `ConstantFolding` and `SimplifyExpression`
