+++ datafusion/src/optimizer/mod.rs
@@ -19,7 +19,6 @@
 //! some simple rules to a logical plan, such as "Projection Push Down" and "Type Coercion".
 
 pub mod common_subexpr_eliminate;
-pub mod constant_folding;
 pub mod eliminate_limit;
 pub mod filter_push_down;
 pub mod limit_push_down;
